+++ react-shopping-cart/src/redux/reducers/index.js
